ria-qa/pyproject.toml
[tool.poetry]
name = "ria-qa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
allure-pytest = "2.8.13"
allure-python-commons = "2.8.13"
attrs = "19.3.0"
certifi = "2020.4.5.1"
chardet = "3.0.4"
colorama = "0.4.3"
configparser = "5.0.0"
crayons = "0.3.0"
Faker = "4.0.3"
future = "0.18.2"
idna = "2.9"
importlib-metadata = "1.6.0"
more-itertools = "8.2.0"
packaging = "20.3"
pluggy = "0.13.1"
py = "1.8.1"
pyparsing = "2.4.7"
pytest = "5.4.1"
pytest-html = "2.1.1"
pytest-metadata = "1.8.0"
python-dateutil = "2.8.1"
requests = "2.23.0"
selene = "1.0.2"
selenium = "3.141.0"
six = "1.14.0"
text-unidecode = "1.3"
urllib3 = "1.25.8"
wcwidth = "0.1.9"
webdriver-manager = "2.3.0"
zipp = "3.1.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
